refactor: attach and subscribe to store for cypress in App.vue
