+++ frontend/src/App.vue
@@ -6,7 +6,31 @@
 
 <script>
 /** @file highest level component of bouldern app */
-export default {};
+
+import {onMounted} from 'vue';
+import {useStore} from 'vuex';
+
+export default {
+  setup() {
+    const store = useStore();
+
+    // Expose Vuex store to cypress tests and attach log hooks when running
+    // cypress test
+    onMounted(() => {
+      if (window.Cypress) {
+        const storeEventHandler = (storeEvent) => {
+          window.Cypress.cy.$log[
+              new Date().toISOString() + ' - ' + storeEvent.type] = {
+            payload: storeEvent.payload,
+          };
+        };
+        store.subscribe(storeEventHandler);
+        store.subscribeAction(storeEventHandler);
+        window['$store'] = store;
+      }
+    });
+  },
+};
 </script>
 
 <style>
